Multiple Device Support via Refactor
