+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "ctap-hid-fido2"
-version = "2.2.4"
+version = "3.0.0"
 authors = ["gebo <35388172+gebogebogebo@users.noreply.github.com>"]
 edition = "2018"
 
@@ -13,7 +13,7 @@ keywords=["FIDO2","CTAP"]
 readme="README.md"
 
 [dependencies]
-aes="0.7.5"
+aes = "0.7.5"
 anyhow = "1.0"
 base64 = "0.13.0"
 block-modes="0.8.1"
