+++ src/fidokey/get_info/get_info_command.rs
