+++ src/lib.rs
@@ -10,9 +10,6 @@ mod bio_enrollment;
 mod bio_enrollment_command;
 pub mod bio_enrollment_params;
 mod bio_enrollment_response;
-mod client_pin;
-mod client_pin_command;
-mod client_pin_response;
 mod config_command;
 mod cose;
 mod credential_management;
@@ -28,10 +25,6 @@ mod get_assertion;
 mod get_assertion_command;
 pub mod get_assertion_params;
 mod get_assertion_response;
-mod get_info;
-mod get_info_command;
-pub mod get_info_params;
-mod get_info_response;
 mod get_next_assertion_command;
 mod hmac;
 mod make_credential;
@@ -54,7 +47,6 @@ pub mod verifier;
 //
 use crate::bio_enrollment_command::SubCommand as BioCmd;
 use crate::bio_enrollment_params::{BioSensorInfo, EnrollStatus1, EnrollStatus2, TemplateInfo};
-use crate::client_pin_command::SubCommand as PinCmd;
 use crate::get_assertion_params::Assertion;
 use crate::get_assertion_params::Extension as Gext;
 use crate::make_credential_params::Attestation;
@@ -72,6 +64,7 @@ mod hid;
 pub type Key = HidParam;
 pub type Cfg = LibCfg;
 
+#[derive(Clone)]
 pub struct LibCfg {
     pub hid_params: Vec<HidParam>,
     pub enable_log: bool,
@@ -92,12 +85,25 @@ impl LibCfg {
 }
 
 /// HID device vendor ID , product ID
-pub struct HidParam {
-    /// vendor ID
-    pub vid: u16,
-    /// product ID
-    pub pid: u16,
+#[derive(Clone)]
+pub enum HidParam {
+    /// Specified when looking for any FIDO device of a certain kind
+    VidPid {vid: u16, pid: u16},
+    Path(String),
 }
+
+/// Struct that contains information about found HID devices. Also
+/// contains a HidParam which can be used to lookup the device
+/// later.
+#[derive(Clone)]
+pub struct HidInfo {
+    pub pid: u16,
+    pub vid: u16,
+    pub product_string: String,
+    pub info: String,
+    pub param: HidParam,
+}
+
 
 impl HidParam {
     /// Generate HID parameters for FIDO key devices
@@ -110,42 +116,15 @@ impl HidParam {
     /// - Nitrokey = vid:0x20a0 , pid:0x42b1
     pub fn get() -> Vec<HidParam> {
         vec![
-            HidParam {
-                vid: 0x1050,
-                pid: 0x0402,
-            }, // yubikey 4/5 u2f
-            HidParam {
-                vid: 0x1050,
-                pid: 0x0407,
-            }, // yubikey 4/5 otp+u2f+ccid
-            HidParam {
-                vid: 0x1050,
-                pid: 0x0120,
-            }, // yubikey touch u2f
-            HidParam {
-                vid: 0x096E,
-                pid: 0x085D,
-            }, // biopass
-            HidParam {
-                vid: 0x096E,
-                pid: 0x0866,
-            }, // all in pass
-            HidParam {
-                vid: 0x0483,
-                pid: 0xa2ca,
-            }, // solokey
-            HidParam {
-                vid: 0x096e,
-                pid: 0x0858,
-            }, // ePass FIDO(A4B)
-            HidParam {
-                vid: 0x20a0,
-                pid: 0x42b1,
-            }, // Nitrokey FIDO2 2.0.0
-            HidParam {
-                vid: 0x32a3,
-                pid: 0x3201,
-            }, // Idem Key
+            HidParam::VidPid { vid: 0x1050, pid: 0x0402 },  // Yubikey 4/5 U2F
+            HidParam::VidPid { vid: 0x1050, pid: 0x0407 },  // Yubikey 4/5 OTP+U2F+CCID
+            HidParam::VidPid { vid: 0x1050, pid: 0x0120 },  // Yubikey Touch U2F
+            HidParam::VidPid { vid: 0x096E, pid: 0x085D },  // Biopass
+            HidParam::VidPid { vid: 0x096E, pid: 0x0866 },  // All in pass
+            HidParam::VidPid { vid: 0x0483, pid: 0xA2CA },  // Solokey 
+            HidParam::VidPid { vid: 0x096E, pid: 0x0858 },  // ePass FIDO(A4B)
+            HidParam::VidPid { vid: 0x20a0, pid: 0x42b1 },  // Nitrokey FIDO2 2.0.0
+            HidParam::VidPid { vid: 0x32a3, pid: 0x3201 },  // Idem Key
         ]
     }
     pub fn auto() -> Vec<HidParam> {
@@ -170,12 +149,12 @@ pub fn hello() {
 }
 
 /// Get HID devices
-pub fn get_hid_devices() -> Vec<(String, HidParam)> {
+pub fn get_hid_devices() -> Vec<HidInfo> {
     hid::get_hid_devices(None)
 }
 
 /// Get HID FIDO devices
-pub fn get_fidokey_devices() -> Vec<(String, HidParam)> {
+pub fn get_fidokey_devices() -> Vec<HidInfo> {
     hid::get_hid_devices(Some(0xf1d0))
 }
 
@@ -183,15 +162,14 @@ fn get_device(cfg: &LibCfg) -> Result<FidoKeyHid> {
     let device = if cfg.hid_params.len() > 0 {
         FidoKeyHid::new(&cfg.hid_params, cfg).map_err(Error::msg)?
     } else {
-        let devs = get_fidokey_devices();
+        let mut devs = get_fidokey_devices();
         if devs.is_empty() {
             return Err(anyhow!("FIDO device not found."));
         }
 
-        let params = vec![HidParam {
-            vid: devs[0].1.vid,
-            pid: devs[0].1.pid,
-        }];
+        let device = devs.pop().unwrap().param;
+
+        let params = vec![device];
         FidoKeyHid::new(&params, cfg).map_err(Error::msg)?
     };
     Ok(device)
@@ -202,66 +180,6 @@ pub fn wink(cfg: &LibCfg) -> Result<()> {
     let device = get_device(cfg)?;
     let cid = ctaphid::ctaphid_init(&device).map_err(Error::msg)?;
     ctaphid::ctaphid_wink(&device, &cid).map_err(Error::msg)
-}
-
-/// Get FIDO key information
-pub fn get_info(cfg: &LibCfg) -> Result<get_info_params::Info> {
-    let device = get_device(cfg)?;
-    get_info::get_info(&device)
-}
-
-/// Get FIDO key information (CTAP 1.0)
-pub fn get_info_u2f(cfg: &LibCfg) -> Result<String> {
-    let device = get_device(cfg)?;
-    get_info::get_info_u2f(&device)
-}
-
-/// Get UV retry count
-pub fn get_uv_retries(cfg: &LibCfg) -> Result<i32> {
-    let device = get_device(cfg)?;
-    let cid = ctaphid::ctaphid_init(&device).map_err(Error::msg)?;
-
-    let send_payload =
-        client_pin_command::create_payload(PinCmd::GetUVRetries).map_err(Error::msg)?;
-
-    let response_cbor = ctaphid::ctaphid_cbor(&device, &cid, &send_payload).map_err(Error::msg)?;
-
-    let pin = client_pin_response::parse_cbor_client_pin_get_retries(&response_cbor)
-        .map_err(Error::msg)?;
-
-    Ok(pin.uv_retries)
-}
-
-/// Get PIN retry count
-pub fn get_pin_retries(cfg: &LibCfg) -> Result<i32> {
-    let device = get_device(cfg)?;
-    let cid = ctaphid::ctaphid_init(&device).map_err(Error::msg)?;
-
-    let send_payload =
-        client_pin_command::create_payload(PinCmd::GetRetries).map_err(Error::msg)?;
-
-    let response_cbor = ctaphid::ctaphid_cbor(&device, &cid, &send_payload).map_err(Error::msg)?;
-
-    let pin = client_pin_response::parse_cbor_client_pin_get_retries(&response_cbor)
-        .map_err(Error::msg)?;
-
-    Ok(pin.retries)
-}
-
-/// Set New PIN
-pub fn set_new_pin(cfg: &LibCfg, pin: &str) -> Result<()> {
-    let device = get_device(cfg)?;
-    let cid = ctaphid::ctaphid_init(&device).map_err(Error::msg)?;
-    client_pin::set_pin(&device, &cid, pin)?;
-    Ok(())
-}
-
-/// Change PIN
-pub fn change_pin(cfg: &LibCfg, current_pin: &str, new_pin: &str) -> Result<()> {
-    let device = get_device(cfg)?;
-    let cid = ctaphid::ctaphid_init(&device).map_err(Error::msg)?;
-    client_pin::change_pin(&device, &cid, current_pin, new_pin)?;
-    Ok(())
 }
 
 fn should_uv(pin: Option<&str>) -> Option<bool> {
@@ -632,81 +550,6 @@ pub fn get_assertion_with_args(cfg: &LibCfg, args: &GetAssertionArgs) -> Result<
     Ok(asss)
 }
 
-#[derive(Debug, Clone, PartialEq)]
-pub enum InfoParam {
-    VersionsU2Fv2,
-    VersionsFido20,
-    VersionsFido21Pre,
-    VersionsFido21,
-    ExtensionsCredProtect,
-    ExtensionsCredBlob,
-    ExtensionsLargeBlobKey,
-    ExtensionsMinPinLength,
-    ExtensionsHmacSecret,
-}
-
-pub fn enable_info_param(cfg: &LibCfg, info_param: &InfoParam) -> Result<bool> {
-    let device = get_device(cfg)?;
-    let info = get_info::get_info(&device)?;
-    let find = match info_param {
-        InfoParam::VersionsU2Fv2 => "U2F_V2",
-        InfoParam::VersionsFido20 => "FIDO_2_0",
-        InfoParam::VersionsFido21Pre => "FIDO_2_1_PRE",
-        InfoParam::VersionsFido21 => "FIDO_2_1",
-        InfoParam::ExtensionsCredProtect => Mext::CredProtect(None).as_ref(),
-        InfoParam::ExtensionsCredBlob => "credBlob",
-        InfoParam::ExtensionsLargeBlobKey => "credBlobKey",
-        InfoParam::ExtensionsMinPinLength => "minPinLength",
-        InfoParam::ExtensionsHmacSecret => Mext::HmacSecret(None).as_ref(),
-    };
-    let ret = info.versions.iter().find(|v| *v == find);
-    if ret.is_some() {
-        return Ok(true);
-    }
-    let ret = info.extensions.iter().find(|v| *v == find);
-    if ret.is_some() {
-        return Ok(true);
-    }
-    Ok(false)
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum InfoOption {
-    Rk,
-    Up,
-    Uv,
-    Plat,
-    ClinetPin,
-    CredentialMgmtPreview,
-    CredMgmt,
-    UserVerificationMgmtPreview,
-    BioEnroll,
-}
-pub fn enable_info_option(cfg: &LibCfg, info_option: &InfoOption) -> Result<Option<bool>> {
-    let device = get_device(cfg)?;
-    let info = get_info::get_info(&device)?;
-    let find = match info_option {
-        InfoOption::Rk => "rk",
-        InfoOption::Up => "up",
-        InfoOption::Uv => "uv",
-        InfoOption::Plat => "plat",
-        InfoOption::ClinetPin => "clientPin",
-        InfoOption::CredentialMgmtPreview => "credentialMgmtPreview",
-        InfoOption::CredMgmt => "credMgmt",
-        InfoOption::UserVerificationMgmtPreview => "userVerificationMgmtPreview",
-        InfoOption::BioEnroll => "bioEnroll",
-    };
-    let ret = info.options.iter().find(|v| (*v).0 == find);
-    if let Some(v) = ret {
-        // v.1 == true or false
-        // - present and set to true.
-        // - present and set to false.
-        return Ok(Some(v.1));
-    }
-    // absent.
-    Ok(None)
-}
-
 /// BioEnrollment - getFingerprintSensorInfo (CTAP 2.1-PRE)
 pub fn bio_enrollment_get_fingerprint_sensor_info(cfg: &LibCfg) -> Result<BioSensorInfo> {
     let device = get_device(cfg)?;
@@ -1043,23 +886,6 @@ pub fn config(cfg: &LibCfg) -> Result<String> {
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn test_client_pin_get_keyagreement() {
-        let hid_params = HidParam::get();
-        let device = FidoKeyHid::new(&hid_params, &Cfg::init()).unwrap();
-        let cid = ctaphid::ctaphid_init(&device).unwrap();
-
-        let send_payload = client_pin_command::create_payload(PinCmd::GetKeyAgreement).unwrap();
-        let response_cbor = ctaphid::ctaphid_cbor(&device, &cid, &send_payload).unwrap();
-
-        let key_agreement =
-            client_pin_response::parse_cbor_client_pin_get_keyagreement(&response_cbor).unwrap();
-        println!("authenticatorClientPIN (0x06) - getKeyAgreement");
-        println!("{}", key_agreement);
-
-        assert!(true);
-    }
 
     #[test]
     fn test_make_credential_with_pin_non_rk_command() {
